Move additional types to registry
